% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precinct_processing_functions.R
\name{convert_toplevel_totals}
\alias{convert_toplevel_totals}
\title{Convert a Single Column of Top-level Totals by Precinct}
\usage{
convert_toplevel_totals(df, column_with_totals, office_text)
}
\arguments{
\item{df}{dataframe with the precinct totals}

\item{column_with_totals}{quoted name of target column containing the totals to capture (e.g. "total_reg_voters")}

\item{office_text}{text to fill in the OpenElex office column as per the standard (e.g. "Registered Voters")}
}
\value{
formatted dataframe ready to match up with OpenElex column structure
}
\description{
This function can be used in situations present in some counties where precinct
totals exist that you want to capture separately, like total ballots cast, total registered voters, etc.
The function assumes the existence of a "precinct" column with the precinct names, and then asks you to
provide the column name (quoted) with the total votes you want to capture. Additionally to match
OpenElex formatting the category is listed as the "office": the third argument allows you to provide
that value.
}
