% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precinct_processing_functions.R
\name{create_infile_string}
\alias{create_infile_string}
\title{Create File Name String for Importing Data}
\usage{
create_infile_string(stateabbr, countyname)
}
\arguments{
\item{stateabbr}{two-letter state abbreviation (e.g. "NY")}

\item{countyname}{county name where the precincts located (e.g. "Saratoga")}
}
\value{
a text string
}
\description{
Can be used to build a file string specific to a county, based on the standardized folder/file format
"NY_Saratoga/NY_Saratoga_GE20_cleaned.xlsx"
}
\examples{
create_infile_string("NY", "Saratoga")
}
