% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precinct_processing_functions.R
\name{create_outfile_string}
\alias{create_outfile_string}
\title{Create File Name String for Saving Final OpenElex File}
\usage{
create_outfile_string(target_state, target_county)
}
\arguments{
\item{target_state}{two-letter state abbreviation}

\item{target_county}{county name}
}
\value{
a file path/name
}
\description{
This function can be used to build a string for the final csv file. It uses the correct naming
convention for the standardized OpenElex names for precinct files.
}
\examples{
create_outfile_string("NY", "Saratoga")
}
