% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigan_specific_functions.R
\name{mi_clean_embedded_precinct_names}
\alias{mi_clean_embedded_precinct_names}
\title{Michigan-specific Function to Fix Mixing of Precinct Names and Vote Types}
\usage{
mi_clean_embedded_precinct_names(data)
}
\arguments{
\item{data}{a dataframe with the above mentioned issues}
}
\value{
returns a cleaned dataframe ready for further processing
}
\description{
A number of Michigan counties report their results in a format that both mixes
together the precinct names and several vote categories (e.g. early vote, total vote)
in a single column. This function both extracts the precinct name from that mess,
and also removes rows featuring  multiple vote subtypes to leave just the "Total" count
for each candidate.
}
\examples{
\dontrun{
mi_clean_embedded_precinct_names(mydataframe)
}
}
