% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigan_specific_functions.R
\name{mi_replace_colnames_presidential}
\alias{mi_replace_colnames_presidential}
\title{Michigan Specific Function to Rename Columns for Presidential Data}
\usage{
mi_replace_colnames_presidential(df)
}
\arguments{
\item{df}{}
}
\value{
df with renamed columns
}
\description{
This function takes a dataframe containing precinct data with the original Michigan headers for presidential
election results and converts the headers to proper formatting to use with package's processing functions. It is tailored
to work with the way Michigan's counties list their candidates/parties and so will only with with that state's results.
}
\examples{
\dontrun{
mi_replace_colnames_presidential(mydataframe)
}
}
