% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigan_specific_functions.R
\name{mi_replace_colnames_ussenate}
\alias{mi_replace_colnames_ussenate}
\title{Michigan Specific Function to Rename Columns for US Senate Dataset}
\usage{
mi_replace_colnames_ussenate(df)
}
\arguments{
\item{df}{}
}
\value{
df with renamed columns
}
\description{
This function takes a dataframe containing precinct data with the original Michigan headers for US Senate
election results and converts the headers to proper formatting to use with package's processing functions
}
\examples{
\dontrun{
mi_replace_colnames_ussenate(mydataframe)
}
}
