% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precinct_processing_functions.R
\name{reshape_precinct_data}
\alias{reshape_precinct_data}
\title{Reshape County Precinct Data for Candidates to OpenElex Format}
\usage{
reshape_precinct_data(df, office, district)
}
\arguments{
\item{df}{formatted dataframe of county precinct-level results}

\item{office}{which office the results refer to (e.g. "presidential", "U.S. House", "State Senate", etc.)}

\item{district}{number for what district is associated with the office (e.g. "45" for the 45th congressional district. note presidential should be left blank as "")}
}
\value{
a reshaped dataframe in openelex long/tidy format
}
\description{
This function is designed to take a wide dataset with one row per precinct and candidate vote
columns and reshape once the initial pre-processing is done.
The file to import should have have a precinct column, followed by columns with vote results formatted as "candidate - party".
The resulting dataframe will be long/tidy data with these columns: precinct, office, district, party, votes.
}
\examples{
\dontrun{
reshape_precinct_data(mydataframe)
}
}
