% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precinct_processing_functions.R
\name{reshape_precinct_data}
\alias{reshape_precinct_data}
\title{Reshaping County Precinct Data to OpenElex Format}
\usage{
reshape_precinct_data(df, office, district)
}
\arguments{
\item{df}{formatted dataframe of county precinct-level results}

\item{office}{which office the results refer to (e.g. presidential, U.S. House, State Senate, etc.)}

\item{district}{number for what district is associated with the office (e.g. "45" for the 45th congressional district. note presidential should be left blank as "")}
}
\value{
a reshaped dataframe in openelex long/tidy format
}
\description{
This function is designed to take a dataset ready to be reshaped once the initial cleanup steps are taken.
The file will have a precinct column, followed by columns with vote results parsed from the "candidate - party" field
}
\examples{
\dontrun{#' reshape_precinct_data(dataframe)}
}
